refactor(e2e): extract variant selector utils from generic CartUtil
Move waitForOptionSelectors and selectDifferentOption into a dedicated
CustomCartMethodUtil class. These methods are only used by the
custom-cart-method recipe tests and don't belong in the generic cart
utility, which caused confusion about whether they were shared helpers.

Addresses fredericoo's review feedback on PR #3546.
